Restore ability to build the jvm side of the library
