+++ core/jvm/src/main/scala/locales/DefaultLocale.scala
@@ -1,0 +1,7 @@
+package locales
+
+import java.util.Locale
+
+object DefaultLocale {
+  def platformLocale: Locale = Locale.getDefault()
+}
