+++ core/shared/src/main/scala/java/text/AttributedCharacterIterator.scala
