+++ core/shared/src/main/scala/java/text/CharacterIterator.scala
