+++ core/shared/src/main/scala/java/text/DecimalFormat.scala
@@ -1,12 +1,12 @@
 package java.text
 
 import java.math.{ RoundingMode, BigDecimal => JavaBigDecimal, BigInteger => JavaBigInteger }
-import java.util.{ Currency, Locale }
-import java.util.LocalesDb
-import locales.{ DecimalFormatUtil, ParsedPattern }
-import scala.math.{ max, min }
 import java.util.concurrent.atomic.AtomicBoolean
 import java.util.concurrent.atomic.AtomicReference
+import java.util.{ Currency, Locale }
+import locales.LocalesDb
+import locales.{ DecimalFormatUtil, ParsedPattern }
+import scala.math.{ max, min }
 
 // The constructor needs a non-localized pattern
 class DecimalFormat(
