+++ core/shared/src/main/scala/locales/cldr/fallback/data/numericsystems.scala
